requirement.txt
alembic==1.10.3
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.3
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
psycopg2==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.10
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1